Rename dependencies package
