+++ dep/configs/lighthouse.go
