+++ dep/configs/lighthouse_validator.go
