+++ dep/configs/prysm_validator.go
