+++ dep/interfaces.go
@@ -1,7 +1,8 @@
-package dependencies
+package dep
 
 import "github.com/urfave/cli/v2"
 
+// UrlBuilder is responsible for parsing a URL for multiple OS options.
 type UrlBuilder interface {
 	ParseUrl(tag, commitHash string) string
 	tag() string
@@ -10,18 +11,22 @@ type UrlBuilder interface {
 	arch() string
 }
 
+// Installer represents a dependency that can be installed and updated.
 type Installer interface {
 	UrlBuilder
 	Install(version string, isUpdate bool) error
 	Update() error
 }
 
+// FileIdentifier ensures that the installed dependency can be located in the file system.
 type FileIdentifier interface {
 	FileName() string
 	FileDir() string
 	FilePath() string
 }
 
+// Client is a generic interface for a client dependency. For a specific layer,
+// use ExecutionClient, ConsensusClient or ValidatorClient
 type Client interface {
 	Installer
 	Start(ctx *cli.Context, arguments []string) error
@@ -35,11 +40,13 @@ type Client interface {
 }
 
 type ExecutionClient interface {
+	Client
 	Peers(ctx *cli.Context) (outbound int, inbound int, err error)
 	Init() error
 }
 
 type ConsensusClient interface {
+	Client
 	Peers(ctx *cli.Context) (outbound int, inbound int, err error)
 }
