+++ flags/logs.go
@@ -3,7 +3,7 @@ package flags
 import (
 	"github.com/urfave/cli/v2"
 
-	"github.com/lukso-network/tools-lukso-cli/dependencies/configs"
+	"github.com/lukso-network/tools-lukso-cli/dep/configs"
 )
 
 var (
